refactor to use a command template runtime package
