+++ pkg/plugin/flavor/swarm/flavor.go
@@ -11,7 +11,7 @@ import (
 	"github.com/docker/go-connections/tlsconfig"
 	"github.com/docker/infrakit/pkg/discovery"
 	group_types "github.com/docker/infrakit/pkg/plugin/group/types"
-	metadata_template "github.com/docker/infrakit/pkg/plugin/metadata/template"
+	runtime "github.com/docker/infrakit/pkg/run/template"
 	"github.com/docker/infrakit/pkg/spi/flavor"
 	"github.com/docker/infrakit/pkg/spi/instance"
 	"github.com/docker/infrakit/pkg/spi/metadata"
@@ -275,21 +275,7 @@ func (s *baseFlavor) prepare(role string, flavorProperties *types.Any, instanceS
 			link:         *link,
 		}
 
-		initTemplate.WithFunctions(func() []template.Function {
-			return []template.Function{
-				{
-					Name: "metadata",
-					Description: []string{
-						"Metadata function takes a path of the form \"plugin_name/path/to/data\"",
-						"and calls GET on the plugin with the path \"path/to/data\".",
-						"It's identical to the CLI command infrakit metadata cat ...",
-					},
-					Func: metadata_template.MetadataFunc(s.plugins),
-				},
-			}
-		})
-		initScript, err = initTemplate.Render(context)
-
+		initScript, err = runtime.StdFunctions(initTemplate, s.plugins).Render(context)
 		log.Debugln(role, ">>> context.retries =", context.retries, "err=", err, "i=", i)
 
 		if err == nil {
